requirements.txt
tensorflow==2.13.0
opencv-python==4.8.0.76
Pillow==9.5.0
imagehash==4.3.1
icrawler==0.6.6
numpy==1.24.3
scikit-learn==1.3.0
